Stop creating webhooks when hit the "Add webhook" button
Creates only after save edited webhook.

+++ includes/admin/settings/views/html-webhooks-edit.php
@@ -143,7 +143,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	<h2><?php esc_html_e( 'Webhook actions', 'woocommerce' ); ?></h2>
 	<table class="form-table">
 		<tbody>
-			<?php if ( '0000-00-00 00:00:00' !== $webhook->get_date_created()->date( 'Y-m-d H:i:s' ) ) : ?>
+			<?php if ( $webhook->get_date_created() && '0000-00-00 00:00:00' !== $webhook->get_date_created()->date( 'Y-m-d H:i:s' ) ) : ?>
 				<?php if ( 0 > $webhook->get_date_modified()->date( 'U' ) ) : ?>
 					<tr valign="top">
 						<th scope="row" class="titledesc">
@@ -172,27 +172,23 @@ if ( ! defined( 'ABSPATH' ) ) {
 					</tr>
 				<?php endif; ?>
 			<?php endif; ?>
-			<?php
-			$delete_url = wp_nonce_url( add_query_arg( array(
-				'delete' => $webhook->get_id(),
-			), admin_url( 'admin.php?page=wc-settings&tab=api&section=webhooks' ) ), 'delete-webhook' );
-			?>
 			<tr valign="top">
 				<td colspan="2" scope="row" style="padding-left: 0;">
 					<p class="submit">
 						<input type="submit" class="button button-primary button-large" name="save" id="publish" accesskey="p" value="<?php esc_attr_e( 'Save webhook', 'woocommerce' ); ?>" />
-						<a style="color: #a00; text-decoration: none; margin-left: 10px;" href="<?php echo esc_url( $delete_url ); ?>"><?php esc_html_e( 'Delete permanently', 'woocommerce' ); ?></a>
+						<?php
+						if ( $webhook->get_id() ) :
+							$delete_url = wp_nonce_url( add_query_arg( array(
+								'delete' => $webhook->get_id(),
+							), admin_url( 'admin.php?page=wc-settings&tab=api&section=webhooks' ) ), 'delete-webhook' );
+							?>
+							<a style="color: #a00; text-decoration: none; margin-left: 10px;" href="<?php echo esc_url( $delete_url ); ?>"><?php esc_html_e( 'Delete permanently', 'woocommerce' ); ?></a>
+						<?php endif; ?>
 					</p>
 				</td>
 			</tr>
 		</tbody>
 	</table>
-</div>
-
-<div id="webhook-logs" class="settings-panel">
-	<h2><?php esc_html_e( 'Webhook logs', 'woocommerce' ); ?></h2>
-
-	<?php // @todo Display webhooks log here. ?>
 </div>
 
 <script type="text/javascript">
